requirements.txt
apache-airflow==2.7.1
spotipy==2.23.0
boto3==1.28.38
psycopg2-binary==2.9.7
dbt-postgres==1.6.0
python-dotenv==1.0.0
pandas==2.0.3
SQLAlchemy==2.0.20 